fix: moved controller-mocks 'package' to utils
